refactor: update palette types
- ensure consistency with canvas api types within the palette-types
  module
- split palette type definitions into seperate files
- remove some stale comments found while review backend
- TG-216
