+++ palette-types/src/PaletteTypes/Rating.ts
@@ -1,0 +1,11 @@
+/**
+ * Represents a rating option within a rubric criterion.
+ */
+
+export interface Rating {
+  id?: number;
+  description: string;
+  longDescription: string;
+  points: number;
+  key: string;
+}
